# frontend/openapi.json
{
 "components": {
  "schemas": {
...
   "DeviceInfoOut": {
    "properties": {
     "area_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Area Id"
     },
     "id": {
      "title": "Id",
      "type": "integer"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "operators": {
      "items": {
       "$ref": "#/components/schemas/DeviceOperatorOut"
      },
      "title": "Operators",
      "type": "array"
     },
     "organization_name": {
      "title": "Organization Name",
      "type": "string"
     },
     "site_id": {
      "title": "Site Id",
      "type": "integer"
     },
     "site_name": {
      "title": "Site Name",
      "type": "string"
     }
    },
    "required": [
     "id",
     "name",
     "site_id",
     "site_name",
     "area_id",
     "organization_name",
     "operators"
    ],
    "title": "DeviceInfoOut",
    "type": "object"
   },
   "DeviceOperatorOut": {
    "properties": {
     "employee_code": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Employee Code"
     },
     "id": {
      "title": "Id",
      "type": "integer"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "pin_length": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Pin Length"
     }
    },
    "required": [
     "id",
     "name",
     "employee_code",
     "pin_length"
    ],
    "title": "DeviceOperatorOut",
    "type": "object"
   },
   "DeviceRegisterIn": {
    "properties": {
     "area_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Area Id"
     },
     "name": {
      "maxLength": 120,
      "minLength": 1,
      "title": "Name",
      "type": "string"
     },
     "site_id": {
      "title": "Site Id",
      "type": "integer"
     }
    },
    "required": [
     "name",
     "site_id"
    ],
    "title": "DeviceRegisterIn",
    "type": "object"
   },
   "DeviceRegisterOut": {
    "properties": {
     "area_id": {
      "anyOf": [
       {
        "type": "integer"
       },
       {
        "type": "null"
       }
      ],
      "title": "Area Id"
     },
     "device_token": {
      "title": "Device Token",
      "type": "string"
     },
     "id": {
      "title": "Id",
      "type": "integer"
     },
     "site_id": {
      "title": "Site Id",
      "type": "integer"
     }
    },
    "required": [
     "id",
     "device_token",
     "site_id",
     "area_id"
    ],
    "title": "DeviceRegisterOut",
    "type": "object"
   },
   "HTTPValidationError": {
    "properties": {
     "detail": {
      "items": {
       "$ref": "#/components/schemas/ValidationError"
      },
      "title": "Detail",
      "type": "array"
     }
    },
    "title": "HTTPValidationError",
    "type": "object"
   },
   "LoginIn": {
    "properties": {
     "email": {
      "maxLength": 254,
      "title": "Email",
      "type": "string"
     },
     "password": {
      "maxLength": 200,
      "minLength": 1,
      "title": "Password",
      "type": "string"
     },
     "totp_code": {
      "anyOf": [
       {
        "maxLength": 10,
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Totp Code"
     }
    },
    "required": [
     "email",
     "password"
    ],
    "title": "LoginIn",
    "type": "object"
   },
   "MeOut": {
    "properties": {
     "email": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Email"
     },
     "employee_code": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Employee Code"
     },
     "id": {
      "title": "Id",
      "type": "integer"
     },
     "language": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Language"
     },
     "name": {
      "title": "Name",
      "type": "string"
     },
     "organization_id": {
      "title": "Organization Id",
      "type": "integer"
     },
     "organization_name": {
      "title": "Organization Name",
      "type": "string"
     },
     "roles": {
      "items": {
       "type": "string"
      },
      "title": "Roles",
      "type": "array"
     },
     "site_ids": {
      "items": {
       "type": "integer"
      },
      "title": "Site Ids",
      "type": "array"
     },
     "totp_enabled": {
      "title": "Totp Enabled",
      "type": "boolean"
     }
    },
    "required": [
     "id",
     "organization_id",
     "name",
     "email",
     "employee_code",
     "roles",
     "site_ids",
     "language",
     "totp_enabled",
     "organization_name"
    ],
    "title": "MeOut",
    "type": "object"
   },
   "Ok": {
    "properties": {
     "ok": {
      "default": true,
      "title": "Ok",
      "type": "boolean"
     }
    },
    "title": "Ok",
    "type": "object"
   },
...
   "PasswordChangeIn": {
    "properties": {
     "current_password": {
      "title": "Current Password",
      "type": "string"
     },
     "new_password": {
      "maxLength": 200,
      "minLength": 10,
      "title": "New Password",
      "type": "string"
     }
    },
    "required": [
     "current_password",
     "new_password"
    ],
    "title": "PasswordChangeIn",
    "type": "object"
   },
   "PinChangeIn": {
    "properties": {
     "pin": {
      "pattern": "^\\d{4,6}$",
      "title": "Pin",
      "type": "string"
     }
    },
    "required": [
     "pin"
    ],
    "title": "PinChangeIn",
    "type": "object"
   },
   "PinLoginIn": {
    "properties": {
     "pin": {
      "pattern": "^\\d{4,6}$",
      "title": "Pin",
      "type": "string"
     },
     "user_id": {
      "title": "User Id",
      "type": "integer"
     }
    },
    "required": [
     "user_id",
     "pin"
    ],
    "title": "PinLoginIn",
    "type": "object"
   },
...
   "SessionOut": {
    "properties": {
     "auth_method": {
      "title": "Auth Method",
      "type": "string"
     },
     "created_at": {
      "format": "date-time",
      "title": "Created At",
      "type": "string"
     },
     "current": {
      "title": "Current",
      "type": "boolean"
     },
     "id": {
      "title": "Id",
      "type": "string"
     },
     "ip": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Ip"
     },
     "last_used_at": {
      "anyOf": [
       {
        "format": "date-time",
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "Last Used At"
     },
     "user_agent": {
      "anyOf": [
       {
        "type": "string"
       },
       {
        "type": "null"
       }
      ],
      "title": "User Agent"
     }
    },
    "required": [
     "id",
     "auth_method",
     "user_agent",
     "ip",
     "created_at",
     "last_used_at",
     "current"
    ],
    "title": "SessionOut",
    "type": "object"
   },
...
   "TokenOut": {
    "properties": {
     "access_token": {
      "title": "Access Token",
      "type": "string"
     },
     "expires_in": {
      "title": "Expires In",
      "type": "integer"
     },
     "token_type": {
      "default": "bearer",
      "title": "Token Type",
      "type": "string"
     },
     "user": {
      "$ref": "#/components/schemas/MeOut"
     }
    },
    "required": [
     "access_token",
     "expires_in",
     "user"
    ],
    "title": "TokenOut",
    "type": "object"
   },
   "TotpCodeIn": {
    "properties": {
     "code": {
      "pattern": "^\\d{6}$",
      "title": "Code",
      "type": "string"
     }
    },
    "required": [
     "code"
    ],
    "title": "TotpCodeIn",
    "type": "object"
   },
   "TotpSetupOut": {
    "properties": {
     "otpauth_uri": {
      "title": "Otpauth Uri",
      "type": "string"
     },
     "secret": {
      "title": "Secret",
      "type": "string"
     }
    },
    "required": [
     "secret",
     "otpauth_uri"
    ],
    "title": "TotpSetupOut",
    "type": "object"
   },
...
   "ValidationError": {
    "properties": {
     "ctx": {
      "title": "Context",
      "type": "object"
     },
     "input": {
      "title": "Input"
     },
     "loc": {
      "items": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "integer"
        }
       ]
      },
      "title": "Location",
      "type": "array"
     },
     "msg": {
      "title": "Message",
      "type": "string"
     },
     "type": {
      "title": "Error Type",
      "type": "string"
     }
    },
    "required": [
     "loc",
     "msg",
     "type"
    ],
    "title": "ValidationError",
    "type": "object"
   }
  },
  "securitySchemes": {
   "HTTPBearer": {
    "scheme": "bearer",
    "type": "http"
   }
  }
 },
 "info": {
  "title": "FloorPulse API",
  "version": "0.1.0"
 },
 "openapi": "3.1.0",
 "paths": {
...
  "/api/v1/auth/device": {
   "get": {
    "description": "Operator picker for a registered shop-floor device.",
    "operationId": "device_info_api_v1_auth_device_get",
    "parameters": [
     {
      "in": "header",
      "name": "x-device-token",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Device-Token"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/DeviceInfoOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Device Info",
    "tags": [
     "auth"
    ]
   }
  },
  "/api/v1/auth/login": {
   "post": {
    "operationId": "login_api_v1_auth_login_post",
    "parameters": [
     {
      "in": "header",
      "name": "user-agent",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "User-Agent"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/LoginIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TokenOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Login",
    "tags": [
     "auth"
    ]
   }
  },
  "/api/v1/auth/logout": {
   "post": {
    "operationId": "logout_api_v1_auth_logout_post",
    "parameters": [
     {
      "in": "header",
      "name": "x-csrf-token",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Csrf-Token"
      }
     },
     {
      "in": "cookie",
      "name": "fp_refresh",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Fp Refresh"
      }
     },
     {
      "in": "cookie",
      "name": "fp_csrf",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Fp Csrf"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Ok"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Logout",
    "tags": [
     "auth"
    ]
   }
  },
  "/api/v1/auth/pin-login": {
   "post": {
    "operationId": "pin_login_api_v1_auth_pin_login_post",
    "parameters": [
     {
      "in": "header",
      "name": "x-device-token",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Device-Token"
      }
     },
     {
      "in": "header",
      "name": "user-agent",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "User-Agent"
      }
     }
    ],
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PinLoginIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TokenOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Pin Login",
    "tags": [
     "auth"
    ]
   }
  },
  "/api/v1/auth/refresh": {
   "post": {
    "operationId": "refresh_api_v1_auth_refresh_post",
    "parameters": [
     {
      "in": "header",
      "name": "x-csrf-token",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "X-Csrf-Token"
      }
     },
     {
      "in": "header",
      "name": "user-agent",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "User-Agent"
      }
     },
     {
      "in": "cookie",
      "name": "fp_refresh",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Fp Refresh"
      }
     },
     {
      "in": "cookie",
      "name": "fp_csrf",
      "required": false,
      "schema": {
       "anyOf": [
        {
         "type": "string"
        },
        {
         "type": "null"
        }
       ],
       "title": "Fp Csrf"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TokenOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "summary": "Refresh",
    "tags": [
     "auth"
    ]
   }
  },
  "/api/v1/devices": {
   "post": {
    "operationId": "register_device_api_v1_devices_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/DeviceRegisterIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "201": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/DeviceRegisterOut"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Register Device",
    "tags": [
     "auth"
    ]
   }
  },
  "/api/v1/devices/{device_id}": {
   "delete": {
    "operationId": "revoke_device_api_v1_devices__device_id__delete",
    "parameters": [
     {
      "in": "path",
      "name": "device_id",
      "required": true,
      "schema": {
       "title": "Device Id",
       "type": "integer"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Ok"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Revoke Device",
    "tags": [
     "auth"
    ]
   }
  },
  "/api/v1/me": {
   "get": {
    "operationId": "me_api_v1_me_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/MeOut"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Me",
    "tags": [
     "auth"
    ]
   }
  },
  "/api/v1/me/password": {
   "post": {
    "operationId": "change_password_api_v1_me_password_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PasswordChangeIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Ok"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Change Password",
    "tags": [
     "auth"
    ]
   }
  },
  "/api/v1/me/pin": {
   "post": {
    "operationId": "change_pin_api_v1_me_pin_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/PinChangeIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Ok"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Change Pin",
    "tags": [
     "auth"
    ]
   }
  },
  "/api/v1/me/sessions": {
   "get": {
    "operationId": "my_sessions_api_v1_me_sessions_get",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "items": {
          "$ref": "#/components/schemas/SessionOut"
         },
         "title": "Response My Sessions Api V1 Me Sessions Get",
         "type": "array"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "My Sessions",
    "tags": [
     "auth"
    ]
   }
  },
  "/api/v1/me/sessions/{family_id}": {
   "delete": {
    "operationId": "revoke_session_api_v1_me_sessions__family_id__delete",
    "parameters": [
     {
      "in": "path",
      "name": "family_id",
      "required": true,
      "schema": {
       "title": "Family Id",
       "type": "string"
      }
     }
    ],
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Ok"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Revoke Session",
    "tags": [
     "auth"
    ]
   }
  },
  "/api/v1/me/totp/disable": {
   "post": {
    "operationId": "totp_disable_api_v1_me_totp_disable_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/TotpCodeIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Ok"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Totp Disable",
    "tags": [
     "auth"
    ]
   }
  },
  "/api/v1/me/totp/enable": {
   "post": {
    "operationId": "totp_enable_api_v1_me_totp_enable_post",
    "requestBody": {
     "content": {
      "application/json": {
       "schema": {
        "$ref": "#/components/schemas/TotpCodeIn"
       }
      }
     },
     "required": true
    },
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/Ok"
        }
       }
      },
      "description": "Successful Response"
     },
     "422": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/HTTPValidationError"
        }
       }
      },
      "description": "Validation Error"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Totp Enable",
    "tags": [
     "auth"
    ]
   }
  },
  "/api/v1/me/totp/setup": {
   "post": {
    "operationId": "totp_setup_api_v1_me_totp_setup_post",
    "responses": {
     "200": {
      "content": {
       "application/json": {
        "schema": {
         "$ref": "#/components/schemas/TotpSetupOut"
        }
       }
      },
      "description": "Successful Response"
     }
    },
    "security": [
     {
      "HTTPBearer": []
     }
    ],
    "summary": "Totp Setup",
    "tags": [
     "auth"
    ]
   }
  },
...
 }
}